# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Invoices API",
        "description": "API de controle financeiro com fluxo de caixa e separação por carteira. Permite lançar contas a pagar e a receber de forma única, parcelada ou mensal.",
        "contact": {
            "name": "My Invoices API Support"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API Base URL"
        }
    ],
    "paths": {
        "/categories": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Listar categorias",
                "description": "Retorna todas as categorias do usuário autenticado",
                "operationId": "0d9f6f774f1b480b5318196b0ce0dc72",
                "responses": {
                    "200": {
                        "description": "Lista de categorias",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Alimentação"
                                            },
                                            "type": {
                                                "type": "string",
                                                "enum": [
                                                    "income",
                                                    "expense"
                                                ],
                                                "example": "expense"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Criar categoria",
                "description": "Cria uma nova categoria para o usuário autenticado",
                "operationId": "d566fee2a2234738bf97d6ef32889c37",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "type"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Alimentação"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "income",
                                            "expense"
                                        ],
                                        "example": "expense"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Categoria criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Alimentação"
                                        },
                                        "type": {
                                            "type": "string",
                                            "enum": [
                                                "income",
                                                "expense"
                                            ],
                                            "example": "expense"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "id": 1,
                                    "name": "Alimentação",
                                    "type": "expense"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro ao criar categoria (ex: categoria duplicada)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Já existe uma categoria com este nome e tipo"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "422": {
                        "description": "Dados de validação inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Dados inválidos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/categories/search": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Buscar categorias",
                "description": "Busca categorias por nome e/ou tipo",
                "operationId": "f41be657140b3b36f1b7d11aab3ca545",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Nome da categoria (busca parcial)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Alimentação"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Tipo da categoria",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "income",
                                "expense"
                            ],
                            "example": "expense"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de categorias encontradas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Alimentação"
                                            },
                                            "type": {
                                                "type": "string",
                                                "enum": [
                                                    "income",
                                                    "expense"
                                                ],
                                                "example": "expense"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados de validação inválidos"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}